requirements.txt
Appium-Flutter-Finder==0.8.0
Appium-Python-Client==5.1.1
astroid==3.3.10
attrs==25.3.0
certifi==2025.6.15
dill==0.4.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
outcome==1.3.0.post0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pylint==3.3.7
PySocks==1.7.1
pytest==8.4.1
selenium==4.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
